requirements.txt
kafka-python-ng
yfinance
pandas
psycopg2-binary
configparser
requests